+++ FeynCalc/Examples/FeynRules/QCDBGF/ValidateModelQCDBGF.m
@@ -25,7 +25,7 @@ If[ $FrontEnd === Null,
 		Print["Validation of the FeynArts model for QCD in background field formalism"];
 ];
 If[$Notebooks === False, $FeynCalcStartupMessages = False];
-$LoadAddOns={"FeynArtsLoader"};
+$LoadAddOns={"FeynArts"};
 <<FeynCalc`
 $FAVerbose = 0;